views/flight: Merge into "flights"
